// elza/elza-react/src/api/generated/model/auth-role.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 *
 * @export
 * @enum {string}
 */
export enum AuthRole {
    CAM1 = 'CAM1',
    CAM2 = 'CAM2',
    CAM3 = 'CAM3',
    CAM4 = 'CAM4',
    CAM5 = 'CAM5',
    CAM6 = 'CAM6'
}